AI-15: moved previous index logic to main to follow the auth pattern laid out in the documentation"
